+++ Project.toml
@@ -2,12 +2,10 @@ name = "JuliaAstroDocs"
 uuid = "5ba9df79-bc07-467f-bade-66a1d49082bd"
 
 [deps]
-CommonMark = "a80b9123-70ca-4bc0-993e-6e3bcb318db6"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 
 [compat]
-CommonMark = "0.9.1"
 SplitApplyCombine = "1.2.3"
 TypedTables = "1.4.6"
 julia = "1.6"
